requirements.txt
ffmpeg-python
openai-whisper
pyannote.audio 
loguru
click
pre-commit